民族scrapy/requirements.txt
alabaster==0.7.10
aliyun-python-sdk-core==2.3.5
aniso8601==3.0.0
asn1crypto==0.24.0
astroid==1.6.3
attrs==17.4.0
Automat==0.6.0
Babel==2.5.3
beautifulsoup4==4.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.2
DBUtils==1.2
docutils==0.14
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
hyperlink==17.3.1
idna==2.6
imagesize==1.0.0
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
Kivy==1.10.0
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.9
kivy.deps.sdl2==0.1.17
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mysql-connector-python==8.0.5
numpy==1.14.3
packaging==17.1
pandas==0.23.1
parsel==1.4.0
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pylint==1.8.4
pylint-flask==0.5
pylint-plugin-utils==0.2.6
PyMySQL==0.7.11
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypiwin32==223
pytesser3==1.0.1
pytesseract==0.2.4
python-dateutil==2.7.3
pytz==2018.4
pywin32==223
PyYAML==3.13
qcloudsms-py==0.1.1
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
Scrapy==1.5.0
selenium==3.11.0
service-identity==17.0.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinx-rtd-theme==0.3.1
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.15
Twisted==18.4.0
urllib3==1.22
visitor==0.1.3
w3lib==1.19.0
Werkzeug==0.14.1
wordcloud==1.4.1
wrapt==1.10.11
WTForms==2.1
zope.interface==4.4.3
